RIB-162: fixed rating calculation and correct rendering

+++ src/main/java/it/rate/webapp/dtos/UserRatingDTO.java
@@ -1,3 +1,3 @@
 package it.rate.webapp.dtos;
 
-public record UserRatingDTO(int rating, String criterionName) {}
+public record UserRatingDTO(Double rating, String criterionName) {}
